+++ instrumentation-genai/opentelemetry-instrumentation-openai-v2/src/opentelemetry/instrumentation/openai_v2/utils.py
@@ -17,7 +17,7 @@ from typing import Mapping, Optional, Union
 from urllib.parse import urlparse
 
 from httpx import URL
-from openai import NOT_GIVEN
+from openai import NotGiven
 
 from opentelemetry._logs import LogRecord
 from opentelemetry.semconv._incubating.attributes import (
@@ -180,7 +180,11 @@ def is_streaming(kwargs):
 
 
 def non_numerical_value_is_set(value: Optional[Union[bool, str]]):
-    return bool(value) and value != NOT_GIVEN
+    return bool(value) and value_is_set(value)
+
+
+def value_is_set(value):
+    return value is not None and not isinstance(value, NotGiven)
 
 
 def get_llm_request_attributes(
@@ -252,10 +256,8 @@ def get_llm_request_attributes(
 
     set_server_address_and_port(client_instance, attributes)
 
-    # filter out None values and NOT_GIVEN values
-    return {
-        k: v for k, v in attributes.items() if v is not None and v != NOT_GIVEN
-    }
+    # filter out values not set
+    return {k: v for k, v in attributes.items() if value_is_set(v)}
 
 
 def handle_span_exception(span, error):