+++ instrumentation-genai/opentelemetry-instrumentation-openai-v2/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3461](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3461))
 - Record prompt and completion events regardless of span sampling decision.
   ([#3226](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3226))
-- Filter out attributes with the value of NOT_GIVEN for chat.completions.create.
+- Filter out attributes with the value of NotGiven instances
   ([#3760](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3760))
 - Migrate off the deprecated events API to use the logs API
   ([#3625](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3628))